static heat-exchanger (no heat transfer calculation, only heating), pressure variation to reach the output condition
The script is going towards a pure heat exchanger calculation (will soon be shifted towards components).
So far the (evaporator) exit temperature of the  working fluid is given and pressure varied such,
that the fluid entering as saturated liquid reaches this point. The mass flow rate ratio is calculated and checked whether the
minimum approach temperature is >0. Other entering states (with quality >0) must be implemented.
The script is not well commented yet.